refactor: extract composer helpers from AgentWorkspace
